Rename image part 2
