+++ common/icons/image.js
