+++ common/views/components/Tasl/Tasl.js
@@ -1,7 +1,8 @@
 // @flow
 
 import {font, spacing} from '../../../utils/classnames';
-import getTaslMarkup from '../../../utils/get-tasl-markup';
+import getLicenseInfo from '../../../utils/get-license-info';
+import {Fragment} from 'react';
 import Icon from '../Icon/Icon';
 
 type Props = {
@@ -14,6 +15,53 @@ type Props = {
   license?: string,
   copyrightHolder?: string,
   copyrightLink?: string
+}
+
+function getMarkup(title, author, sourceName, sourceLink, license, copyrightHolder, copyrightLink) {
+  const licenseInfo = license && getLicenseInfo(license);
+  console.log(getTitleHtml(title, author, sourceLink));
+  return (
+    <Fragment>
+      {getTitleHtml(title, author, sourceLink)}
+      {getSourceHtml(sourceName, sourceLink)}
+      {getCopyrightHtml(copyrightHolder, copyrightLink)}
+      {licenseInfo &&
+        <a rel='license' href={licenseInfo.url}>{licenseInfo.text}</a>
+      }
+    </Fragment>
+  );
+}
+
+function getTitleHtml(title, author, sourceLink) {
+  if (!title) return;
+
+  const byAuthor = author ? `, ${author}` : '';
+
+  if (sourceLink) {
+    return <Fragment><a href={sourceLink} property='dc:title' rel='cc:attributionURL'>{title}{byAuthor}</a>. </Fragment>;
+  } else {
+    return <Fragment><span property='dc:title'>{title}{byAuthor}.</span> </Fragment>;
+  }
+}
+
+function getSourceHtml(sourceName, sourceLink) {
+  if (!sourceName) return '';
+
+  if (sourceLink) {
+    return <Fragment>Source: <a href={sourceLink} rel='cc:attributionURL'>{sourceName}</a>. </Fragment>;
+  } else {
+    return <Fragment>Source: {sourceName}. </Fragment>;
+  }
+}
+
+function getCopyrightHtml(copyrightHolder, copyrightLink) {
+  if (!copyrightHolder) return '';
+
+  if (copyrightLink) {
+    return <Fragment>&copy; <a href={copyrightLink}>{copyrightHolder}</a>. </Fragment>;
+  } else {
+    return <Fragment>&copy; {copyrightHolder}. </Fragment>;
+  }
 }
 
 const Tasl = ({isFull, contentUrl, title, author, sourceName, sourceLink, license, copyrightHolder, copyrightLink}: Props) => (
@@ -38,8 +86,8 @@ const Tasl = ({isFull, contentUrl, title, author, sourceName, sourceLink, licens
     <div className={`
       drawer__body js-show-hide-drawer bg-black font-white
       ${spacing({s: 1}, {padding: ['top', 'bottom', 'left']})}
-      ${spacing({s: 6}, {padding: ['right']})}`}
-    dangerouslySetInnerHTML={{__html: getTaslMarkup({title, author, sourceName, sourceLink, license, copyrightHolder, copyrightLink})}}>
+      ${spacing({s: 6}, {padding: ['right']})}`}>
+      {getMarkup(title, author, sourceName, sourceLink, license, copyrightHolder, copyrightLink)}
     </div>
     {isFull &&
       <button className="tasl__button absolute plain-button js-show-hide-trigger">
